refactor: moved models to dto and adjusted models

+++ telehealth-fe/src/main/java/org/drpl/telefe/model/Medicine.java
@@ -1,26 +1,30 @@
 package org.drpl.telefe.model;
 
+import java.math.BigDecimal;
+
 public class Medicine {
     private Long id;
     private String name;
     private String description;
     private String dosage;
-    private double price;
+    private BigDecimal price;
 
-    // Diperlukan konstruktor default untuk deserialisasi Jackson
     public Medicine() {}
 
-    // Getters and Setters
     public Long getId() { return id; }
     public void setId(Long id) { this.id = id; }
+
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
+
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
+
     public String getDosage() { return dosage; }
     public void setDosage(String dosage) { this.dosage = dosage; }
-    public double getPrice() { return price; }
-    public void setPrice(double price) { this.price = price; }
+
+    public BigDecimal getPrice() { return price; }
+    public void setPrice(BigDecimal price) { this.price = price; }
 
     @Override
     public String toString() {
